Fix Tutorial.md and README.md code (#274)
* Fix tutorial code to match text

Spread the previous state into the next state returned by the reducer.
Even though this redux state only has one property, `counter`, keeping
the other state is good to do in general. Also, this keeps the code
consistent. Currently, the code sometimes grabs the whole `state` and
other times destructures just `{ counter }` from the `state`. This fix
always grabs the whole `state` to be able to spread it.

The current code for `createActions()` for `'INCREMENT'` and
`'DECREMENT'` that accept an argument always returns a `payload`
of `{ amount: 1 }` and `{ amount: -1 }` respectively. I am assuming the
intent is to default to 1 or -1. This fix uses a default parameter
and matches the example in README.md.

The `+ amount` was changed into `- amount` for the `'DECREMENT'` reducer
in a previous commit. `+ amount` is correct because `amount` is negative
for the `'DECREMENT'` action. It also makes sense like this because
below, `combineActions()` is used meaning the `'INCREMENT'` and
`'DECREMENT'` reducers should be the same.

* Update README.md code to match Tutorial.md